+++ commands/mentor/getServerRoles.js
@@ -1,5 +1,8 @@
 module.exports = (message) => {
+  // creates an empty list
   let serverRoles = [];
+
+  // gets all roles from a messages' guild;
   message.guild.roles.fetch()
     .then(roles => {
       serverRoles.push(roles);
