+++ commands/mentor/reactToMessage.js
@@ -1,0 +1,4 @@
+module.exports = (message, emojiList) => {
+  // reacts to message with every emoji from emojiList
+  emojiList.forEach(emoji => message.react(emoji));
+}
